Bug#35478833 : set persist authentication_ldap_simple_bind_root_pwd                is not saving password
Description:
Variable value of authentication_ldap_simple_bind_root_pwd is
data masked. This prevents the real variable value from being
persisted when set persist is used.

Persisting the sensitive variable creates a new entry without removing
the previous entry of the variable in the configuration file.

The server does not start if persisted sensitive variables
cannot be interpreted even if the server option
persist_sensitive_variables_in_plaintext is set to ON.
As documented the server should start with a warning in this case.

Fix:
Variable value data masking of authentication_ldap_simple_bind_root_pwd
removed. Variable authentication_ldap_simple_bind_root_pwd
made a sensitive variable.

A check that removes the previous entry when the new entry of the
sensitive variable is persisted in the configuration file.

Allow server start up with a warning if applicable when the
server option persist_sensitive_variables_in_plaintext is set to ON.

Change-Id: I504669793cae89042f3d3928d5999cc5cecc7256