Bug#35642087 All secondaries crashes when there is a network inactivity around 20 seconds on the primary host
Problem
===================

When a primary host experiences a network inactivity
around 20 seconds,all secondaries fail in a small chance.

Failure happens in
Gcs_xcom_communication_protocol_changer::decrement_nr_packets_in_transit()

Analysis
===================

sender: node that wants to send messages
leader: node in the group that is able to propose messages
        to the group.
        Currently it could be 1 or all.
channel: stream of proposals flowing in the group.

The failure happens in
Gcs_xcom_communication_protocol_changer::decrement_nr_packets_in_transit
when we try to search for a node_id from the ORIGINATOR
node that does not exist on the
Gcs_xcom_nodes const &xcom_nodes list.

If we backtrack to where this list and the originator
synode are built, we end up on the entry point for messages
in GCS coming from XCom, which is cb_xcom_receive_data
function in gcs_xcom_interface.cc, where all seems correct.

Diving into XCom itself, specifically in
xcom/xcom_interface.cc::deliver_to_app we notice that,
right before calling xcom_receive_data, we do create the
origin node, from data that is contained in the user data
from the accepted group message.

This field is filled by a sender in
xcom_base.cc::proposer_task, right after requesting a synode
from the leader if the group is running with all nodes NOT
being a leader. It is done in
brand_client_msg(ep->client_msg->p, ep->msgno) on the
xcom_base.cc::proposer_task loop.

Mind again that the xcom_base.cc::proposer_task task
looping, runs solely on the sender of the message.

This is needed to enforce that, when in Paxos Single Leader
is enabled, we actually know who sent the message, because,
since we only have one channel, all message synode origin
would be from the leader on that configuration.

In very rare occasions, a message can be proposed but not
approved, the sender faces an issue and it is removed from
the group, and someone else picks up that proposal and gets
it approved on behalf of the crashed node, as per the PAXOS
algorithm.

This can lead to one of two scenarios:
origin, the message that was supposed to be in XCom
Configuration C, ends up being approved on a configuration
C+1, where the sender no longer exists.

origin, the message that was supposed to be in XCom
Configuration C, ends up being approved on a configuration
C+1, where the sender that had a node_id
K, now is a K', which is fact a different node, because
node_id can vary between configuration C and C+1.

To solve this, we will always check the synode brand
vs. the actual message brand to analyze if this message
was intended to a previous configuration, but it was only
delivered on the current configuration, thus providing the
CORRECT site to everyone that needs an origin.

Change-Id: If9c3922448e9010b1dc7e9aae67a0d827b8d00a5