Bug#36724336  BLOB_INLINE_SIZE=0 is ignored
When BLOB_INLINE_SIZE=0 is used as a NDB_COLUMN option for column's
COMMENT, the option is ignored and the default inline size for the
BLOB type is used.

The main issue is that the function that reads the conversion was
also ignoring a successful conversion to 0 ( <= 0 ).

Changed to use strtol for better conversion functionality and made
sanity check just use "< 0".

Change-Id: Ieb652f720c626443c66a7ff1303c1e7f4fe283f1