Bug#36925906 External stored routines should always use utf8mb4 charset/collation
External stored routines will always use utf8mb4. This means that
even if the connection that creates a routine is using a different
charset/collation, any SQL call-out within the routine
will use utf8mb4.

We currently set the client charset to utfmb4, but we do not do the
same for the connection collation.

When creating a non-SQL routine, this patch sets the
collation_connection column in the routines table to
utf8mb4_0900_ai_ci (the default utf8mb4 collation) if original
collation_connection is not utf8mb4.  If the collation_connection is a
utf8mb4 collation, we keep the original collation.

Change-Id: Id8f489eba843179e713c4d1ec450d7e891cb1970