Bug# 36816986 - MySQL Shell command injection
Problem:
mysqldump does not sanitize the SQL statements
which might contain some special characters like
new line feed etc.

Because of this, a user can inject arbitrary shell
command(s) in to the dump file created by mysqldump.

And if that file is executed on a host machine, the
shell command will get executed on that host machine.

Solution:
The fix is to make sure that the rows which are being
fetched from dump_tablespaces are sanitized properly.

This fix is just sanitizing the rows coming in from
mysql_fetch_row while dumping tablespaces.

Change-Id: Ide74636ba60fd3cb6c936149f76b3c4ead70ecc9