Bug#36861370 UBSAN failures in NDB unit tests
Second undefined behaviour problem where in SimpleProperties an unaligned
char array is potentially referenced
as Uint32 and this is treated as undefined behaviour by UBSAN. Fix by
renaming the putWords()
function to putWordsFromChars() and make it take a char array  pointer
as first argument, which will then be
written as words to the properties buffer. Note that caller of
putWordsFromChars() still need to make sure that all bytes in the last
word are properly zero terminated.

Change-Id: I8dbab28c8a4fced1de69d40593969eb839b81893