Bug#36845105 Massive regression in query performance for a query having 12 joins
This bug was a regression due to the fix for Bug#36125903 "Hypergraph:
Inconsistent numeric precision for AccessPath in optimizer trace"
(Chng-Id: I44a04c7ec78e494f7fb988e5b3e32104af3844a1).

A new regular expression was created for every number that was formatted
when generating optimizer trace for access paths. That is expensive.
This commit instead uses my_fcvt_internal() to format numbers.

In addition it:
* Fixes an error that could give mantissas with four digits, e.g.
  "1000e+3" rather than "1e+6".

* Eliminates a call to std::log10().

* Adds microbenchmarks and unit tests.

Change-Id: I3f575c5ffa6cdbc2d29cae37de7e4f08591f5f45