Bug#36734772: Hypergraph optimizer fails to use index for t1 INNER JOIN t2 RIGHT JOIN t3
The hypergraph optimizer did not use indexes in some queries where an
inner join was on the inner side of an outer join or on the outer side
of a semijoin.

This happened if both tables in the inner join could use index
lookups, and one of the index lookups was parameterized on the other
table in the inner join and the other index lookup was parameterized
on the table on the other side of the outer join or semijoin.

The problem was that DisallowParameterizedJoinPath() was too
restrictive, and rejected all good plans for the query. It tries to
avoid creating parameterized join paths if the parameterization could
have been resolved earlier, so that the optimizer doesn't have to
spend time unnecessarily on plans that are unlikely to be any better
than the alternatives. In these cases, the parameterization couldn't
have been resolved earlier, so it is too strict to reject these plans.

This code is now changed so that it also allows creating a
parameterized join if the left side of the join fully resolves the
parameterization of the right side.

Change-Id: I20728c24fe20128a76ae96772255a648842e65c3