Bug#36861370 UBSAN failures in NDB unit tests [noclose]
First undefined behaviour problem where the LogHandler keeps track of the
last used log level to make sure that log message is always printed when
log level changes. The last log level value is however initialized to a
value not part of the enum and thus it's undefined behaviour. Fix by
adding a new enum value to be used for initalization.

Change-Id: Ibe4eff27e84bd762fdf5f8120a98d870ff23c6be