Bug#36739383 Assertion 'false' materialize_iterator::SpillState::read_next_row_secondary_over [follow-up]
Seen in some of the HW tests. This is a case of a too large row for
the default hash table space, which next also overflows the normal tmp
table (error 135 HA_ERR_RECORD_FILE_FULL) and must therefore move to
an InnoDB tmp table. The logic didn't handle the case of a single row
too large for the hash table properly in this case.

Change-Id: I71f3ad884f576179ca03b20abb5fa398858b51d9