Bug#36855521: main.derived_condition_pushdown segfaults with hypergraph
A query in main.derived_condition_pushdown occasionally failed when
running with the hypergraph optimizer.

The problem was that one of the input tables to a hash join was an
internally generated temporary table, and the hash join code was not
prepared for this possibility. In such tables,
table->pos_in_table_list is nullptr, so the hash join code shouldn't
unconditionally dereference table->pos_in_table_list.

Fixed by adding checks for nullptr.

Change-Id: I299933c95a1b1330f821ee7f2d62f1b01b98cd47