Bug#36796197: BIT query wrong result
Aggregation function SUM yields different result for the same value if
DISTINCT is also used. Reason why this happens is, when using DISTINCT
a temporary table is used to hold the values so that in the end
only the unique values are returned to the user. When creating such
table while setting up DISTINCT, we should take into consideration that
Item_sum_sum already figured the data type & length and use that
since we want the same data type instead of the current approach
which is to recalculate it, resulting in different data types.

Change-Id: I6de4385d96400a0d52c09ab151c4a454b8bc4664