Bug#36560156: simple query crash in count_field_types() -> Item_ref::real_item()
The preparation of table value constructors doesn't keep track of the
number of hidden columns added for expressions in the ORDER BY clause.
This could cause problems later in the resolving and make the server
exit.

Fixed by counting the number of hidden items in table value
constructors the same way as it's done in regular query blocks.

Change-Id: Icb62214afb9a4a56c3bb92c648c1460cbf4b82c0