Bug#36892945 - MLE crash if set signal and variable are set
When a user variable is used in a DML statement, the variable and
its value are allocated in the THD::user_var_events_alloc. Before
the DML statement is written to the binlog, the user variable and
its corresponding values from THD::user_var_events_alloc are also
recorded in binlog.

For stored procedures(SP), memory for user variables and their
associated values is allocated within the procedure's execution
memory root as each DML statement is written separately to the
binlog. At the end of stored procedure execution, THD::user_var_events_alloc
is set to nullptr and it is reset to THD::mem_root before executing
the next statement.

However, when executing statements using the Statement handle
interface, THD::user_var_events_alloc is not reset to THD::mem_root
after statement execution or before processing the next statement.
This leads to the reported issue when any statement is executed
following a stored procedure using the Statement handle interface.

To address this issue, THD::user_var_events_alloc is now reset to the
statement execution memory root after executing a sub-statement from
the statement handle interface.

Change-Id: Ie943d36e6f75dfa673e8d9ca5afd9125d21533a9