Bug#36967379 Upgrade to libcurl 8.9.1 [patches]
Additional patch, disabling the creation of an ALIAS library called CURL::libcurl

This change
-CMAKE_MINIMUM_REQUIRED(VERSION 3.5.1)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.18)
triggered:
  ADD_LIBRARY cannot create ALIAS target "CURL::libcurl" because another
  target with the same name already exists.

release notes for 3.18 say:

Creation of an ALIAS target overwriting an existing target now raises
an error. See policy CMP0107.
The ``OLD`` behavior for this policy is to allow target overwrite.
The ``NEW`` behavior of this policy is to prevent target overwriting.

Change-Id: If60a067440030d0236359803f753c8b235032aad