Bug#36978299: Rename some local variables in transform_table_subquery_to_derived
This is a preparation for the work in WL#13052, about extending support
for transformation into derived tables also for general quantified
comparison predicates.

When dealing with a subquery and its outer context, it is customary to
refer to the "inner" objects (concerning the subquery) and the "outer"
objects (concerning the parent query block). It is also customary to
refer to a variable pointing to a query block with the abbreviation qb
and a variable pointing to a query expression with the abbreviation qe.

Hence, the actions that will be performed are:

- Change name of subs_query_expression to inner_qe.
- Change name of subs_query_block to inner_qb.

Also, change name of argument subq to subq_pred, to indicate that it is
a predicate involving a (table) subquery.

Change-Id: Ibcf09080da405d18cf573eed4d6c90225da28166