Bug#36981486 routertest_component_routing_sharing fails in show_warnings_without_server_connection
The test sporadically fails on a slower machines with

  Lost connection

when it should just succeed.

Background
==========

When the test closes all server-side connections to check
that SHOW WARNINGS in the router works with out a server-side connection,
the server-side connection of the current client-connection
may still be attached to the client.

If it is attached, closing the server-side connection will also close
the client-side connection.

The test should wait until the server-side connection is properly
stashed before closing it as that will then NOT lead to a close of the
client connection (as it is expected).

Change
======

- wait until the client's server connection is stashed before closing
  it.

Change-Id: Ia1e9102604e72813a25039106189cb8ffbe8c486