Bug#36980184: Use first-row-count instead of init-cost as cost dimension in hypergraph [1/2, refactor, noclose]
Create a helper function that calculates the cost of reading the first
row from an access path, so that code that needs to calculate the
first row cost doesn't need to do the calculation manually. Use this
function where the first row cost is currently calculated.

The existing calculation in CreateIteratorFromAccessPath() was wrong,
because it included the init cost also in the per-row cost. This is
corrected by the patch.

Change-Id: I79e6f38dbcae08c4b476eb07a2e901136fb188a0