Bug#36980184: Use first-row-count instead of init-cost as cost dimension in hypergraph [2/2, compare]
Make CompareAccessPaths() use first_row_cost() instead of init_cost()
as cost dimension.

When init_cost() is a cost dimension, the hypergraph optimizer will
keep the candidates that are the cheapest ones to initialize and read
zero rows from. A better dimension would be the cost of initializing
the path and reading the amount of rows needed to reach the LIMIT in
the query. Currently, there is no estimate for the number of rows that
have to be read from a path to reach the LIMIT. But it's reasonable to
expect that more often than not, at least one row will be read. So
using first_row_cost() will in most cases be slightly better than
using init_cost().

This change could give better plans for some queries using LIMIT.
Another effect is that it could speed up planning time, because more
candidates with a very high total cost and zero init cost can be
rejected earlier when a fraction of the very high total cost is taken
into the value. This is seen to reduce planning time in some complex
queries.

Change-Id: Ida4fcf734b0e5973010891bb97ee0423c9ff7840