WL#16396 Change default value for innodb_log_writer_threads based on available cores
Description:
------------
  The innodb_log_writer_threads determines whether dedicated threads are
used to flush InnoDB log files to the disk. When disabled, the current
thread which generated the log files will flush them to the disk.
If enabled, dedicated log writer threads will be spawned and they will
handle the flushing operation.

Improvement:
------------
  This commit changes the default value of innodb_log_writer_threads
based on the number of logical CPUs. It is disabled if the number of
logical CPUs is less than 32 and enabled otherwise.

Change-Id: Iddb36a7fc6705e268c49e37c74bb52a1debdb610