Bug#36989337 NDB Varbinary like columns fail with OO_SETVALUE
Problem:

When writing values to varbinary columns using NdbRecord and OO_SETVALUE
in the NDBAPI the write will fail with error 829 "Corrupt data received
for insert/update". This since the value when sent to NDB cluster will
be wrapped within a too small attribute which is checked in DBTUP in
data nodes.

The bug affects columns that uses the NDB Varchar, Varbinary,
Longvarchar, and, Longvarbinary types.

Fix:

Make sure that the attribute will have room for both the bare value and
the extra length bytes.

Test:

A new test is added to demonstrate failure:

  testNdbApi -n SetVarbinaryWithSetValue WIDE_COL

Change-Id: I1dc2885457c59e7f19ea71637240504d57fd69a1