WL#16099: Support statement modifiers in the hypergraph optimizer
Statement modifier support in hypergraph:

1) Add support for the QUICK modifier for DELETE statements.
The old optimizer handles single-table DELETEs differently from
multi-table DELETEs. The logic to handle the QUICK statement
modifier is present in the single-table DELETE path. Since
hypergraph handles both single-table DELETES and multitable
DELETES using the multi-table path, it never hits the QUICK
handling code, so the QUICK modifier is ignored by hypergraph.

This is fixed by adding a check in hypergraph to detect single-
table DELETEs to pass the HA_EXTRA_QUICK flag to the engine if
the QUICK modifier is present in the query.

2) Fix an incorrect result in SQL_CALC_FOUND_ROWS.
For SELECT queries with SQL_CALC_FOUND_ROWS and LIMIT, the flag
to count all the rows should be set while processing the LIMIT.
If the query contains a DISTINCT, the count_all_rows flag is set
for the DISTINCT access path. This results in the rowcount being
calculated wrt the DISTINCT, ignoring the GROUP BY on top of the
DISTINCT.

This is fixed by ensuring that the count_all_rows is never set
while processing DISTINCT. The flag will only be set at the
topmost LIMIT.

The remaining statement modifiers are supported in hypergraph by
the existing code in the old optimizer.

Test cases:

1) Add test coverage for SQL_BUFFER_RESULT and SQL_BIG_RESULT to
check that temporary tables are created or avoided as per the
modifier.

2) Add test case for SELECT HIGH_PRIORITY to check locking
priority wrt UPDATE statements.

Existing test cases provide test coverage for the remaining
statement modifiers.

Change-Id: Ie0f3ab2983d7375c0548a08487fb01c6eaacc868