Bug#37006524 	InnoDB: Assertion failure: buf0flu.cc:621:block->page.in_LRU_list
Symptom:
During recovery, on successful page read callback after applying hashed redo entries to that page it is added to the flush list with `buf_flush_insert_sorted_into_flush_list()` which asserts the page is in LRU and this assertion fails.

Root cause:
The `in_LRU_list` flag is protected by the `buf_pool->LRU_list_mutex` and the `buf_flush_insert_sorted_into_flush_list()`  does not hold it when checking it is set.
It was earlier identified that holding block->mutex is enough to make sure the block will not get evicted to be removed from the LRU. But this flag will get set to false while holding the LRU mutex also in case of `buf_LRU_make_block_young()` which calls `buf_LRU_remove_block()` that resets the flag and `buf_LRU_add_block_low()` that sets it back again.

Fix:
We can't check this condition in assertion without the LRU mutex, and as this is a debug-only assertion and the subsequent check for `in_page_hash` should also assure the page is not evicted we will remove the assertion.

Change-Id: Ie241f39c903df60e87a0f07f7334b3bc3e709b20