Bug#36775910: Record buffer not set in index range scans
Post-push fix:
It was found that enabling record buffer for range scans
caused performance regression in sysbench tests for
higher thread count. Different buffer sizes were tried
and it seems to be that smaller buffer sizes benefit
small range queries while the bigger buffer benefitted
the large ranges and table scans. So the solution is to
to pull the buffer size towards a minimum buffer size
if the range is not that large, while still pulling
the buffer size towards the maximum buffer size for
large ranges and table scans. To acheive this, a fraction
(1/10th) of the estimated rows is used to determine
the rows to be fit in the record buffer. After this
adjustment, if the required size falls out of the range
defined, it is adjusted to the range closer to the value.
The MAX_RECORD_BUFFER_SIZE remains same i.e 128KB.
MIN_RECORD_BUFFER_SIZE is newly defined to be 4KB.
If all the estimated rows for a scan can fit in a
smaller buffer than the MIN_RECORD_BUFFER_SIZE, only
the required size is allocated.

Change-Id: I208a76fa6993b136288c5419f3bdcd4d1fe34b04