Post-push fix for Bug#36072163 and Bug#36072349
The sanity checks added recently to start_mysqld.inc found that the test
innodb.import_compress_encrypt tried to start_mysqld.inc while it was
already up and running.
This was a previously unnoticed consequence of a patch for
WL#15938 Remove keyring_file plugin
which commented out the portion of the script which shutdown_mysqld.inc,
but kept the part which then sourced start_mysqld.inc.

This patch resolves the problem by commenting the start_mysqld.inc as well.

Change-Id: Ic24d6a42007406234a8bbd64d171624a0909e92a