Bug#28550140 FIX RECONNECTTEST IN CLUSTERJ
This bug report, from 2018, notes that on rare occasions the test
ReconnectTest will fail and report "Retry count too low: 0"; the
problem report says "the retry count is too low due to a race
condition in the threads run by the test."

In my analysis, I found that a test thread obtains a lock on the
retryCount when incrementing it (line 339; line 358; line 371)
but that the main thread neglects to acquire this lock when reading
the retryCount at line 197.

Replace retryCount with an AtomicInteger, and reenable the test.

This patch is for 7.5 and up.

Change-Id: I62563be7d94f7431f90855a1cf717a24991124bc