Bug#37014535: Assertion path->type == AccessPath::GROUP_INDEX_SKIP_SCAN || (path->type == AccessPath::FILTER && path->filter().child [1/2, refactor, noclose]
Preparatory refactoring before the fix for the bug.

A helper function ForEachChild() is factored out of WalkAccessPaths(),
so that it can be used instead of WalkAccessPaths() in cases where we
want to visit only the immediate children of a given access path.

WalkAccessPaths() is changed to use this new helper function to avoid
code duplication.

Change-Id: Id6e4b2aa4cef8b946b397fd730bbc4955c9ec00b