Bug#36235324 testSystemRestart -n Bug27434 T1, error 5046 is never hit
Problem

Error insert 5046 is left after testcase runs.

Solution

Seems like error insert 5046 is/should not be reachable as it
is in a branch which would normally ndbassert(false), therefore
it could only act as a way to make a crash handleable in a testcase.

Since there is no other special action to make this branch reachable
that seems unlikely.

Rather it seems like the error insert 5045 should be used here,
causing a node failure during the SR takeover, and giving coverage
of the code to handle that case.

Therefore error insert 5046 is removed, and the test is modified
to use 5045 which results in the test passing, and no error insertion
being left behind.

Change-Id: Ic7887caac1761be0be26ef1e2e5fb8345a0aab18