BUG#37015554: Fix metadata handling in PARTITION UNLOAD DDL
When all partitions of a table are unloaded from secondary engine, the
table level secondary_load flag is also unset indicating the table is no
longer loaded in the secondary engine. The code compares list of already
loaded partitions and list of partitions getting unloaded in the DDL to
decide whether to reset table level flag or not. However, the existing
logic was incorrectly comparing parent level partitions with leaf level
partitions thus incorrectly resetting table's secondary_load flag when a
sub-partition is unloaded after loading a parent partition. The problem
is not seen when the table has no sub-partitions.

In this commit, we fixed the logic to compare only leaf level partitions
which gives consistent list irrespective of whether table has
sub-partitions or not.

Change-Id: I9c183748f85fa548adff6b0404232c0a0c3e5b3e