Bug#36804785 `false' in operator at sql/sql_executor.cc
A const item that is part of group by and not found in
select list is not added as a hidden item to the fields
list post the fix for Bug#34951115. This is taken into
consideration while checking for replacements for
expressions involving ROLLUP operation during temp table
creation i.e If a const item is not found the field list,
same item is used as replacement as part of Bug#36444257.
However, that fix only checked for const_item() and not
const_for_execution(). The current bug reported fails
the assertion for items that are not marked const, but
only marked const for execution.
Fix is to now check for const_for_execution() instead
of const_item().

Change-Id: I35c717ec23010c69852623f336adfb77818f5080