Bug#33664699: hypergraph: Nested loop cross joins look too cheap compared to hash cross joins Bug#33967325: Hypergraph chooses slow nested loop joins over hash joins
The queries in these bug reports now use hash joins when using the
hypergraph optimizer, and the performance is on par with the old
optimizer.

This was fixed partly by WL#16107, which improved the cost
calculation, and partly by bug#36625968, which made the hypergraph
optimizer prefer hash joins heuristically when there is no index that
could help speed up a nested loop join.

Adjust tests that previously had workarounds for this bug.

Change-Id: I0a523afd8dcee1313f4466dcc06b0e248debe07f