Bug#36982911 mysql_harness::join does not handle std::string_view
mysql_harness::join() concats the elements of a container into a string.

It doesn't support joining if elements are std::string_view,
nor if the container is a c-array or std::initializer_list.

Also, benchmarking showed, that the use of std::accumulate() is slow in
this case as it leads to copying the joined std::string in each round
instead of appending to one string.

Joining 1024 std::string_views into a comma-seperated string:

join() via std::accumulate()    11310 ns/iter
join() via direct loop           7043 ns/iter

Change
======

- simplified the implementation into a generic implementation which
  relies on std::begin() and avoids std::accumulate()
- added support to join strings-views and c-arrays
- added tests for string-views, c-arrays and initializer-lists
- added benchmarks

Change-Id: I3285b8470d85495c09cc51b1574d91173d38c099