Bug#35854686 - Server crash on function JOIN::update_depend_map(ORDER*)+0x98
Issue:
Server is crashing for a specific query consisting of subqueries
and a group by clause.

Analysis:
When optimising a group by clause for a query block, where the
group by field is a part of subquery, the optimizer was trying to
refer a table which belongs to a different query block causing
the server to crash.

Fix:
Fixed the above issue by checking if the field belongs to the
outer query block.

Change-Id: I65e1170f3aad7f3566ff55a61916c211f2a20456