Bug#36867806 Heap-use-after-free: regexp::Regexp_facade::~Regexp_facade()
An SQL statement had heap-use-after-free when run as a prepared statement.
The failing statemen was a functional index in a create table statement,
and objects created ended up in the execution MEM_ROOT.

The solution is to switch to the prepare MEM_ROOT when doing resolving
of functional indexes.

Change-Id: Ic54ef5eaaa4b90e1018cb82b9c73f5f6c5e29e01
(cherry picked from commit 2db31275b840b02c5b970da1caa5468f905514e1)