Post-push fix for Bug#36302624 performance_schema.data_lock_waits is generated with O(N^3) complexity
Problem:

On 32-bit systems alignof(uint64_t) might be 4.
This in turn permits alignof(lock_t) == 4 and sizeof(lock_t)%8 == 4.
Taken together, this means that the bitmap which is immediately after
the lock_t struct can have an address which is not a multiple of 8.
This is suboptimal as Bitmap::find_set(..) handles such case by entering
a slow path. There is a debug-only assert which checks that the address
of bitmap is a multiple of 8 to ensure high performance.
This assertion is failing on 32-bit builds.

Solution:

Added alignas(8) to definition of lock_t to enforce proper alignment.
Added a static assert that alignof(lock_t) is at least 8, to prevent
regression.
Added a static assert that heap allocator respects alignof(lock_t).
Added an assert to dict_mem_table_create that the lock_t allocated there
for autoinc is also properly aligned.
As sizeof(T) must be at least alignof(T), this also ensures that
sizeof(lock_t) is a multiple of 8.

Note: the instances of lock_t are obtained in various ways:

a) By calling ut::malloc_withkey(...,REC_LOCK_SIZE * REC_LOCK_CACHE) to
allocate one big chunk, and then slicing it into REC_LOCK_CACHE items
of REC_LOCK_SIZE bytes. This is fine, because allocator ensures
alignof(std::max_align_t) which must be at least that of double. So, the
address of the big buffer is at least divisible by 8. And because
REC_LOCK_SIZE is defined as sizeof(ib_lock_t) + 256, so as long as
alignof(lock_t) is at least 8, then adding 256 to it will not spoil the
property that each smaller buffer has an address divisible by 8.

b) Similar to a) but with TABLE_LOCK_SIZE and TABLE_LOCK_CACHE. Here the
difference is that TABLE_LOCK_SIZE doesn't include 256 bytes of padding
for the bitmap, which makes the proof even simpler. Also we don't ever
use bitmaps for table locks, so this shouldn't matter in practice.

c) Calls to mem_heap_alloc in RecLock::lock_alloc, lock_table_create and
dict_mem_table_create. Each of them should give an address which has at
least UNIV_MEM_ALIGNMENT==8, so should be fine. There are asserts in
each of these places (this patch adds the missing one) that the
allocated lock object has at least alignof(lock_t).

This is why we expect all the lock_t structs to have at least
alignof(lock_t) and why it is sufficient to ensure alignof(lock_t) is at
least 8.

Change-Id: I7d4892d28baefd11aa1c0fa4ac2dba502f270bd2